requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
grpcio==1.66.1
grpcio-tools==1.66.1
httpx==0.27.2
pydantic==2.9.2
